Struct rpds::queue::Queue

pub struct Queue<T, P = RcK>
where
    P: SharedPointerKind,
{ /* private fields */ }

A persistent queue with structural sharing.

Complexity

Let n be the number of elements in the queue.

Temporal complexity

Operation	Average	Worst case

new()	Θ(1)	Θ(1)

enqueue()	Θ(1)	Θ(1)

dequeue()	Θ(1)	Θ(n)

dequeue() amortized	Θ(1)	Θ(1)

peek()	Θ(1)	Θ(1)

len()	Θ(1)	Θ(1)

clone()	Θ(1)	Θ(1)

iterator creation	Θ(1)	Θ(1)

iterator step	Θ(1)	Θ(n)

iterator full	Θ(n)	Θ(n)

Implementation details

This queue is implemented as described in Immutability in C# Part Four: An Immutable Queue.

Implementations

impl<T> Queue<T, ArcK>

pub fn new_sync() -> QueueSync<T>

impl<T> Queue<T>

pub fn new() -> Queue<T>

impl<T, P> Queue<T, P>
where P: SharedPointerKind,

pub fn new_with_ptr_kind() -> Queue<T, P>

pub fn peek(&self) -> Option<&T>

pub fn dequeue(&self) -> Option<Queue<T, P>>

pub fn dequeue_mut(&mut self) -> bool

pub fn enqueue(&self, v: T) -> Queue<T, P>

pub fn enqueue_mut(&mut self, v: T)

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn iter(&self) -> Iter<'_, T, P>