Struct rpds::vector::Vector

pub struct Vector<T> { /* fields omitted */ }

A persistent vector with structural sharing.

Complexity

Let n be the number of elements in the vector.

Temporal complexity

Operation 	 Best case 	 Average 	 Worst case

new() 	 Θ(1) 	 Θ(1) 	 Θ(1)

set() 	 Θ(log(n)) 	 Θ(log(n)) 	 Θ(log(n))

push_back() 	 Θ(log(n)) 	 Θ(log(n)) 	 Θ(log(n))

drop_last() 	 Θ(log(n)) 	 Θ(log(n)) 	 Θ(log(n))

first()/last()/get() 	 Θ(log(n)) 	 Θ(log(n)) 	 Θ(log(n))

len() 	 Θ(1) 	 Θ(1) 	 Θ(1)

clone() 	 Θ(1) 	 Θ(1) 	 Θ(1)

iterator creation 	 Θ(1) 	 Θ(1) 	 Θ(1)

iterator step 	 Θ(1) 	 Θ(1) 	 Θ(log(n))

iterator full 	 Θ(n) 	 Θ(n) 	 Θ(n)

Proof sketch of the complexity of full iteration

A tree of size n and degree d has height ⌈logd(n)⌉ - 1.

A complete iteration is a depth-first search on the tree.

A depth-first search has complexity Θ(|V| + |E|), where |V| is the number of nodes and |E| the number of edges.

The number of nodes |V| for a complete tree of height h is the sum of powers of d, which is (dʰ - 1) / (d - 1). See Calculating sum of consecutive powers of a number.

The number of edges is exactly |V| - 1.

By 2. and 3. we have that the complexity of a full iteration is

     Θ(|V| + |E|)
   = Θ((dʰ - 1) / (d - 1))      (by 4. and 5.)
   = Θ(dʰ)
   = Θ(n)                       (by 1.)

Implementation details

This vector is implemented as described in Understanding Persistent Vector Part 1 and Understanding Persistent Vector Part 2.

Methods

impl<T> Vector<T>

pub fn new() -> Vector<T>

pub fn new_with_bits(bits: u8) -> Vector<T>

pub fn first(&self) -> Option<&T>

pub fn last(&self) -> Option<&T>

pub fn get(&self, index: usize) -> Option<&T>

pub fn set(&self, index: usize, v: T) -> Option<Vector<T>>

pub fn set_mut(&mut self, index: usize, v: T) -> bool

Returns true if the operation was successful.

pub fn push_back(&self, v: T) -> Vector<T>

pub fn push_back_mut(&mut self, v: T)

pub fn drop_last(&self) -> Option<Vector<T>>

pub fn drop_last_mut(&mut self) -> bool

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn iter(&self) -> Iter<T>